pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "provider-selection-agent"
version = "0.1.0"
description = "CLI agent for comparing service providers with deterministic scoring and LLM-assisted recommendation drafts."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.60",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "pydantic>=2.7",
    "python-dotenv>=1.0",
    "fastapi>=0.115",
    "typer>=0.12",
    "streamlit>=1.44",
    "pandas>=2.2",
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "rich>=13.7",
    "tenacity>=8.3",
    "chromadb>=0.5",
    "beautifulsoup4>=4.12",
    "duckduckgo-search>=7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.6",
    "mypy>=1.10",
    "types-PyYAML>=6.0",
]

[project.scripts]
compare-providers = "provider_selection_agent.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.mypy]
python_version = "3.11"
disallow_untyped_defs = true
check_untyped_defs = true
no_implicit_optional = true
ignore_missing_imports = true
warn_return_any = false
warn_unused_ignores = true
exclude = ["tests"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-q -p no:cacheprovider"
